# backend/openapi.json
{"openapi":"3.1.0","info":{"title":"VigilAero Backend","version":"0.2.0"},"paths":{"/auth/login":{"post":{"tags":["auth"],"summary":"Login","operationId":"login_auth_login_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/threats/assign_operator":{"post":{"tags":["threats"],"summary":"Assign Operator","operationId":"assign_operator_api_threats_assign_operator_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/DroneAssignmentRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"HTTPBearer":[]}]}},"/api/threats/templates":{"get":{"tags":["threats"],"summary":"Get Templates","operationId":"get_templates_api_threats_templates_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/threats/run":{"post":{"tags":["threats"],"summary":"Run Threat","operationId":"run_threat_api_threats_run_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ThreatRunRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"HTTPBearer":[]}]}},"/incidents/active":{"get":{"tags":["incidents"],"summary":"Get Active Incidents","operationId":"get_active_incidents_incidents_active_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}},"security":[{"HTTPBearer":[]}]}},"/incidents/{incident_id}/mitigate":{"post":{"tags":["incidents"],"summary":"Mitigate","operationId":"mitigate_incidents__incident_id__mitigate_post","security":[{"HTTPBearer":[]}],"parameters":[{"name":"incident_id","in":"path","required":true,"schema":{"type":"string","title":"Incident Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/MitigationRequest"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/incidents/{incident_id}/close":{"post":{"tags":["incidents"],"summary":"Close","operationId":"close_incidents__incident_id__close_post","security":[{"HTTPBearer":[]}],"parameters":[{"name":"incident_id","in":"path","required":true,"schema":{"type":"string","title":"Incident Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/api/threats/assign_operator":{"post":{"tags":["threats"],"summary":"Assign Operator","operationId":"assign_operator_api_api_threats_assign_operator_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/DroneAssignmentRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"HTTPBearer":[]}]}},"/api/api/threats/templates":{"get":{"tags":["threats"],"summary":"Get Templates","operationId":"get_templates_api_api_threats_templates_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/api/threats/run":{"post":{"tags":["threats"],"summary":"Run Threat","operationId":"run_threat_api_api_threats_run_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ThreatRunRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"HTTPBearer":[]}]}},"/api/incidents/active":{"get":{"tags":["incidents"],"summary":"Get Active Incidents","operationId":"get_active_incidents_api_incidents_active_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}},"security":[{"HTTPBearer":[]}]}},"/api/incidents/{incident_id}/mitigate":{"post":{"tags":["incidents"],"summary":"Mitigate","operationId":"mitigate_api_incidents__incident_id__mitigate_post","security":[{"HTTPBearer":[]}],"parameters":[{"name":"incident_id","in":"path","required":true,"schema":{"type":"string","title":"Incident Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/MitigationRequest"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/incidents/{incident_id}/close":{"post":{"tags":["incidents"],"summary":"Close","operationId":"close_api_incidents__incident_id__close_post","security":[{"HTTPBearer":[]}],"parameters":[{"name":"incident_id","in":"path","required":true,"schema":{"type":"string","title":"Incident Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/security/zerotrust":{"get":{"tags":["security"],"summary":"Get Zerotrust","operationId":"get_zerotrust_security_zerotrust_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}},"security":[{"HTTPBearer":[]}]},"post":{"tags":["security"],"summary":"Set Zerotrust","operationId":"set_zerotrust_security_zerotrust_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ZeroTrustUpdate"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"HTTPBearer":[]}]}},"/security/events":{"get":{"tags":["security"],"summary":"Security Events","operationId":"security_events_security_events_get","security":[{"HTTPBearer":[]}],"parameters":[{"name":"drone_id","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Drone Id"}},{"name":"limit","in":"query","required":false,"schema":{"type":"integer","maximum":500,"minimum":1,"default":200,"title":"Limit"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/security/zerotrust":{"get":{"tags":["security"],"summary":"Get Zerotrust","operationId":"get_zerotrust_api_security_zerotrust_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}},"security":[{"HTTPBearer":[]}]},"post":{"tags":["security"],"summary":"Set Zerotrust","operationId":"set_zerotrust_api_security_zerotrust_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ZeroTrustUpdate"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"HTTPBearer":[]}]}},"/api/security/events":{"get":{"tags":["security"],"summary":"Security Events","operationId":"security_events_api_security_events_get","security":[{"HTTPBearer":[]}],"parameters":[{"name":"drone_id","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Drone Id"}},{"name":"limit","in":"query","required":false,"schema":{"type":"integer","maximum":500,"minimum":1,"default":200,"title":"Limit"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/forensics":{"get":{"tags":["forensics"],"summary":"Get Forensics","operationId":"get_forensics_api_forensics_get","security":[{"HTTPBearer":[]}],"parameters":[{"name":"drone_id","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Drone Id"}},{"name":"incident_id","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Incident Id"}},{"name":"limit","in":"query","required":false,"schema":{"type":"integer","maximum":500,"minimum":1,"default":200,"title":"Limit"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/incident/respond":{"post":{"tags":["incidents"],"summary":"Respond To Incident","operationId":"respond_to_incident_api_incident_respond_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/IncidentRespondRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"HTTPBearer":[]}]}}},"components":{"schemas":{"DroneAssignmentRequest":{"properties":{"drone_id":{"type":"string","title":"Drone Id"},"operator":{"type":"string","title":"Operator"}},"type":"object","required":["drone_id","operator"],"title":"DroneAssignmentRequest"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"IncidentRespondRequest":{"properties":{"drone_id":{"type":"string","title":"Drone Id"},"response_id":{"type":"string","title":"Response Id"},"response_name":{"type":"string","title":"Response Name"}},"type":"object","required":["drone_id","response_id","response_name"],"title":"IncidentRespondRequest"},"LoginRequest":{"properties":{"username":{"type":"string","title":"Username"},"password":{"type":"string","title":"Password"}},"type":"object","required":["username","password"],"title":"LoginRequest"},"LoginResponse":{"properties":{"access_token":{"type":"string","title":"Access Token"},"token_type":{"type":"string","title":"Token Type","default":"bearer"},"role":{"type":"string","title":"Role"},"expires_at":{"type":"string","title":"Expires At"}},"type":"object","required":["access_token","role","expires_at"],"title":"LoginResponse"},"MitigationRequest":{"properties":{"action":{"type":"string","title":"Action"}},"type":"object","required":["action"],"title":"MitigationRequest"},"ThreatRunRequest":{"properties":{"company_id":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Company Id"},"drone_id":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Drone Id"},"threat_type":{"type":"string","title":"Threat Type"},"training":{"type":"boolean","title":"Training","default":true}},"type":"object","required":["threat_type"],"title":"ThreatRunRequest"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"},"ZeroTrustUpdate":{"properties":{"policy":{"type":"string","title":"Policy"}},"type":"object","required":["policy"],"title":"ZeroTrustUpdate"}},"securitySchemes":{"HTTPBearer":{"type":"http","scheme":"bearer"}}}}